Allow reuse of locally built wheels
As of this commit we use a custom version of pip.

+++ src/pypi2nix/stage3.py
@@ -55,7 +55,7 @@ let
               (builtins.attrValues pkgs)}; do
             if [ -d "$dep/bin" ]; then
               for prog in "$dep/bin/"*; do
-                if [ -f $prog ]; then
+                if [ -x "$prog" ] && [ -f "$prog" ]; then
                   ln -s $prog $out/bin/`basename $prog`
                 fi
               done
